assignment list moved to components in sections
