+++ Web Ui/src/sections/Assignments/components/AssignmentsList.tsx
@@ -15,13 +15,13 @@ import {
   Button,
 } from "@mui/material";
 import { styled } from "@mui/system";
-import { AssignmentDataObject } from "../../modules/Assignments/domain/assignmentInterfaces";// Import your assignment model
+import { AssignmentDataObject } from "../../../modules/Assignments/domain/assignmentInterfaces";// Import your assignment model
 
-import { fetchAssignmentsUseCase } from "../../application/views/assignmentsViews/useCases/fetchAssignmentsApater"; // Import your fetchAssignments function\
-import { deleteAssignmentUseCase } from "../../application/views/assignmentsViews/useCases/deleteAssignmentAdapter";
-import { sendAssignemtUseCase } from "../../application/views/assignmentsViews/useCases/sendAssignmentsAdapter";
-import { ConfirmationDialog } from "./ConfirmationDialog";
-import { GitLinkDialog } from "./GitHubLinkDialog";
+import { fetchAssignmentsUseCase } from "../../../application/views/assignmentsViews/useCases/fetchAssignmentsApater"; // Import your fetchAssignments function\
+import { deleteAssignmentUseCase } from "../../../application/views/assignmentsViews/useCases/deleteAssignmentAdapter";
+import { sendAssignemtUseCase } from "../../../application/views/assignmentsViews/useCases/sendAssignmentsAdapter";
+import { ConfirmationDialog } from "../ConfirmationDialog";
+import { GitLinkDialog } from "../GitHubLinkDialog";
 
 
 const ButtonContainer = styled("div")({
